requirements.txt
requests==2.31.0
websockets==12.0
pandas==2.1.4
numpy==1.26.2
SQLAlchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
pytest==7.4.3
pydantic==2.5.2
